feat: simple base class add mock executor
